Move dummyData
dummydata 폴더를 만들어서 더미 데이터들을 모아두려 함
