+++ app/frontend/src/component/mainpage/MainPage.tsx
@@ -3,7 +3,7 @@ import { ModalController, ChatContent, ConfigContent } from '../modal/Modal'
 import NavBar from './navbar/NavBar'
 import '/src/scss/MainPage.scss'
 import EasyFetch from './../../utils/EasyFetch';
-import { testFriendList } from './dummyData'
+import { testFriendList } from '../../dummydata/testFriendList'
 
 /*!
  * @author yochoi, donglee
