+++ app/frontend/src/dummydata/testFriendList.ts
